migrate off OldException
This is a rather tricky change.  Many exceptions that used to be caught will no longer be, but in general I believe most of the intended exceptions are IO.  Also, behavior of "fail" within Either String seems to have changed.  I've made many Monad m functions less generic to try to simplify things.