Fix horrible bug causing only unverifiable puffs to be displayed
While recovering from a fat-finger that resulted in darcs happily blowing away
all the crypto-related changes I'd committed, I accidentally restored a WIP
version of Gale/KeyCache.hs and committed it instead of the final version.  The
binary I'd most recently built was one I'd ensured worked correctly, and I
thought I was restoring the same code that corresponded to that binary, but
unfortunately, this was not the case.

I apologize for this unacceptable mistake.